Enum leo_errors::cli::cli_errors::CliError

pub enum CliError {
    FormattedError(FormattedError),
    BacktracedError(BacktracedError),
}

CliError enum that represents all the errors for the leo-lang crate.

Variants

FormattedError(FormattedError)

BacktracedError(BacktracedError)

Implementations

impl CliError

pub fn opt_args_error(error: impl ErrorArg) -> Self

Not actually ever returned anywhere outside a test.

pub fn unidentified_api() -> Self

For when APM returns an unidentifed API error.

pub fn unable_to_connect_aleo_pm() -> Self

For when the CLI is unable to connect to a Leo Package Manager.

pub fn package_not_found() -> Self

For when APM is unable to find to a package.

pub fn unkown_api_error(status: impl Display) -> Self

For when APM returns an unknown API error.

pub fn account_not_found() -> Self

For when the APM account username is not registered, or password is incorrect.

pub fn incorrect_password() -> Self

For when APM account password is incorrect.

pub fn bad_request(msg: impl Display) -> Self

For when a request to APM fails witha bad request status.

pub fn not_logged_in() -> Self

For when the user is performing some APM action that requires the user to be logged in first.

pub fn already_published() -> Self

For when a package with the same name and author name is already published.

pub fn internal_server_error() -> Self

For when APM is experiencing a HTTP Status of 500.

pub fn reqwest_json_error(error: impl ErrorArg) -> Self

For when the reqwest library fails to get the request JSON.

pub fn incorrect_command_argument() -> Self

For when the user provides an incorrect command argument.

pub fn cli_io_error(error: impl ErrorArg) -> Self

For when the CLI experiences an IO error.

pub fn cli_bytes_conversion_error(error: impl ErrorArg) -> Self

For when the CLI experiences a bytes conversion error.

pub fn cli_zip_error(error: impl ErrorArg) -> Self

For when the CLI experiences a zip error.

pub fn manifest_file_not_found() -> Self

For when the CLI cannot find the manifest file.

pub fn unable_to_get_user_token() -> Self

For when the CLI was unable to get the user token.

pub fn supplied_token_is_incorrect() -> Self

For when the CLI was supplied an incorrect user token.

pub fn stored_credentials_expired() -> Self

For when the CLI user’s stored credentials expire.

pub fn no_credentials_provided() -> Self

For when the user does not provide credentials to the CLI.

pub fn logout_permision_denied() -> Self

For when the CLI does not have persmission to modify the .leo folder and cannot logout the user.

pub fn cannot_access_logout_file() -> Self

For when the CLI cannot access the logout file.

pub fn package_cannot_be_named_after_a_keyword() -> Self

For when the user tries to name a package after a Leo keyword.

pub fn no_package_description() -> Self

For when the user has not provided a package description.

pub fn missing_package_license() -> Self

For when the user has not provided a package license.

pub fn missing_package_remote() -> Self

For when the package is missing its remote section in the Leo.toml file.

pub fn package_author_is_not_set() -> Self

For when the user has not provided the package author field in the Leo.toml file.

pub fn failed_to_convert_to_toml(error: impl ErrorArg) -> Self

For when the CLI fails to convert an object to TOML.

pub fn failed_to_convert_from_toml(error: impl ErrorArg) -> Self

For when the CLI fails to TOML an object.

pub fn package_directory_does_not_exist() -> Self

For when the current package directory doesn’t exist.

pub fn invalid_project_name() -> Self

For when the current project has an invalid name.

pub fn invalid_package_name(name: impl Display) -> Self

For when the current package has an invalid name.

pub fn package_main_file_not_found() -> Self

For when the package main.leo file is not found.

pub fn package_directory_already_exists(path: impl Debug) -> Self

For when the package directory already exists.

pub fn package_could_not_create_directory(error: impl ErrorArg) -> Self

For when the CLI could not a directory.

pub fn unable_to_setup() -> Self

For when the CLI could not setup a Leo command.

pub fn program_file_does_not_exist(path: impl Display) -> Self

For when the program file does not exist.

pub fn could_not_fetch_versions(error: impl ErrorArg) -> Self

For when the CLI could not fetch the versions.

pub fn unable_to_watch(error: impl ErrorArg) -> Self

For when the CLI failed to watch the Leo package.

pub fn failed_to_enable_ansi_support() -> Self

For when the CLI fails to enable ansi support.

pub fn self_update_error(error: impl ErrorArg) -> Self

For when the CLI fails to self update.

pub fn self_update_build_error(error: impl ErrorArg) -> Self

For when the CLI fails to self update.

pub fn old_release_version(current: impl Display, latest: impl Display) -> Self

For when the CLI has an old release version.

pub fn dependencies_are_not_installed() -> Self

pub fn recursive_dependency_found(message: impl Display) -> Self

pub fn unable_to_read_imported_dependency_manifest() -> Self

pub fn num_exit_codes() -> i32

impl CliError

pub fn remove_token_and_username(error: Error) -> Self

Possible errors that can be thrown when the CLI is removing the token and username.