leo_errors
Module common

Contains the common functionalities for defining errors.

Re-exports

pub use self::backtraced::*;

pub use self::formatted::*;

pub use self::traits::*;

Modules

backtraced

This module contains a backtraced error and its methods.

formatted

This module contains a formatted error and its methods.

macros

This module contains the macros for making errors easily.

traits

This module contains traits for making errors easily.