leo_errors::common

Module formatted

This module contains a formatted error and its methods.

Structs

Formatted

Formatted compiler error type undefined value x –> file.leo: 2:8 | 2 | let a = x; | ^ | = help: Initialize a variable x first. Makes use of the same fields as a BacktracedError.

Label

Represents error labels.

Enums

Color

Represents available colors for error labels.

Functions

compute_line_spans 🔒

Compute the start and end columns of the highlight for each line

print_code_line 🔒

Print a single line of code with connector and optional highlight

print_gap 🔒

Print a gap or ellipsis between blocks of code

print_multiline_underline 🔒

Print the final underline for a multi-line highlight (Rust-style with -)