Trait leo_errors::common::traits::LeoErrorCode

pub trait LeoErrorCode: Sized {
    fn exit_code(&self) -> i32;

    fn error_code(&self) -> String;

    fn exit_code_mask() -> i32;

    fn error_type() -> String;

    fn code_identifier() -> i8 { ... }
}

ErrorCode trait that all Errors should implement.

Required methods

fn exit_code(&self) -> i32

Returns the error’s exit code for the program.

fn error_code(&self) -> String

Returns the prefixed error identifier.

fn exit_code_mask() -> i32

Returns the error’s exit code mask, as to avoid conflicts.

fn error_type() -> String

Returns the error’s code type for the program.

Provided methods

fn code_identifier() -> i8

The LeoErrorCode which has a default code identifier of 037 (Leo upsidedown and backwards). This is to make the exit codes unique to Leo itself.