Enum leo_errors::compiler::compiler_errors::CompilerError

pub enum CompilerError {
    FormattedError(FormattedError),
    BacktracedError(BacktracedError),
}

CompilerError enum that represents all the errors for the leo-compiler crate.

Variants

FormattedError(FormattedError)

Tuple Fields of FormattedError

0: FormattedError

BacktracedError(BacktracedError)

Tuple Fields of BacktracedError

0: BacktracedError

Implementations

impl CompilerError

pub fn invalid_test_context(name: impl Display) -> Self

For when the test function has invalid test context.

pub fn file_read_error(path: impl Debug, error: impl ErrorArg) -> Self

For when the compiler can’t read a file from the provided path.

pub fn no_main_function() -> Self

For when there is no main function in a Leo program.

pub fn no_test_input() -> Self

For when the compiler can’t find the test input files with the specified name.

pub fn console_fmt_expected_left_or_right_brace(span: &Span) -> Self

For when the console formatter expected a left or right brace after a left brace.

pub fn console_fmt_expected_escaped_right_brace(span: &Span) -> Self

For when the console formatter expected a right brace after a right brace.

pub fn console_container_parameter_length_mismatch(
containers: impl Display,
parameters: impl Display,
span: &Span
) -> Self

For when the amount of arguments, and number of containers mismatch in a format statement.

pub fn console_assertion_depends_on_input(span: &Span) -> Self

For when a user tries to user console assert outside of a test function.

pub fn console_assertion_failed(span: &Span) -> Self

For when a console assert fails.

pub fn console_assertion_must_be_boolean(span: &Span) -> Self

For when a console assert value is not a boolean.

pub fn cannot_enforce_expression(
operation: impl Display,
error: impl ErrorArg,
span: &Span
) -> Self

For when a experssion gadget oepration cannot be enforced due to a SnarkVM syntehsis error.

pub fn cannot_evaluate_expression(operation: impl Display, span: &Span) -> Self

For when an expression has mismatching types for an operation.

pub fn array_length_out_of_bounds(span: &Span) -> Self

For when an array length goes out of the legal bounds of 2^32.

pub fn array_index_out_of_legal_bounds(span: &Span) -> Self

For when an array index goes out of the legal bounds of 2^32.

pub fn conditional_boolean_expression_fails_to_resolve_to_bool(
actual: impl Display,
span: &Span
) -> Self

For when a boolean expression does not resolve to a boolean type.

pub fn expected_circuit_member(expected: impl Display, span: &Span) -> Self

For when the expected circuit member could not be found.

pub fn incompatible_types(operation: impl Display, span: &Span) -> Self

For when the operation has no implementation on the type of variable received.

pub fn tuple_index_out_of_bounds(index: impl Display, span: &Span) -> Self

For when a tuple index goes out of the tuples size bounds.

pub fn array_index_out_of_bounds(index: impl Display, span: &Span) -> Self

For when an array index goes out of the arrays size bounds.

pub fn array_invalid_slice_length(span: &Span) -> Self

For when a invalid array slice length is requested.

pub fn invalid_index_expression(actual: impl Display, span: &Span) -> Self

For when an array index does not resolve to an integer type.

pub fn unexpected_array_length(
expected: impl Display,
actual: impl Display,
span: &Span
) -> Self

For when a typed array receives an assignment of an array with a different length.

pub fn invalid_circuit_static_member_access(
member: impl Display,
span: &Span
) -> Self

For when the circuit static member does not exist.

pub fn undefined_array(actual: impl Display, span: &Span) -> Self

For when a user is trying to use an array in access expression before it is declared.

pub fn undefined_circuit(actual: impl Display, span: &Span) -> Self

For when the user is trying to us a circuit that is not yet declared.

pub fn undefined_identifier(name: impl Display, span: &Span) -> Self

For when the user tries to use an identifier not declared in scope.

pub fn undefined_circuit_member_access(
circuit: impl Display,
member: impl Display,
span: &Span
) -> Self

For when the user tries to access an undefined circuit member.

pub fn input_variable_type_mismatch(
expected: impl Display,
actual: impl Display,
variable: impl Display,
span: &Span
) -> Self

For when the input variable type mismatches the declared function input type.

pub fn expected_const_input_variable(
variable: impl Display,
span: &Span
) -> Self

For when the declared function input variable was expected to be constant but the input file did not have it in the constants section.

pub fn expected_non_const_input_variable(
variable: impl Display,
span: &Span
) -> Self

For when the declared function input variable was expected to be mutable but the input file did not have it in the main section.

pub fn invalid_function_input_array(actual: impl Display, span: &Span) -> Self

For when the declared function input variable was expected to be a valid array in the input file.

pub fn invalid_input_array_dimensions(
expected: impl Display,
actual: impl Display,
span: &Span
) -> Self

For when the declared function input variable was expected to be an array with differing dimensions.

pub fn input_tuple_size_mismatch(
expected: impl Display,
actual: impl Display,
span: &Span
) -> Self

For when the declared function input variable was expected to be a tuple with a different number of arguments.

pub fn invalid_function_input_tuple(actual: impl Display, span: &Span) -> Self

For when the declared function input variable was expected to be a valid tuple in the input file.

pub fn function_input_not_found(
function: impl Display,
expected: impl Display,
span: &Span
) -> Self

For when the declared function input variable was expected to be a valid tuple in the input file.

pub fn double_input_declaration(input_name: impl Display, span: &Span) -> Self

For when the declared function input variable was defined twice in the input file.

pub fn output_not_enough_registers(span: &Span) -> Self

For when the input file does not define enough registers.

pub fn output_mismatched_types(
left: impl Display,
right: impl Display,
span: &Span
) -> Self

For when the input file register types do not match the output types being generated.

pub fn output_file_io_error(error: impl ErrorArg) -> Self

For when there’s an IO error with the output file.

pub fn output_file_cannot_read(path: impl Debug) -> Self

For when the output file cannot be read.

pub fn output_file_cannot_remove(path: impl Debug) -> Self

For when the output file cannot be removed.

pub fn statement_array_assign_index(span: &Span) -> Self

For when the user tries to index a single array more than once.

pub fn statement_array_assign_index_const(span: &Span) -> Self

For when the user tries to use a non const value as an index.

pub fn statement_array_assign_interior_index(span: &Span) -> Self

For when the user tries to assign an index to something not an array of length >= 1;

pub fn statement_array_assign_range(span: &Span) -> Self

For when the user tries to assign a range of values to something that expected a single value.

pub fn statement_array_assign_index_bounds(
index: impl Display,
length: impl Display,
span: &Span
) -> Self

For when the user tries to index a value from an array that is >= the array length.

pub fn statement_array_assign_range_order(
start: impl Display,
stop: impl Display,
length: impl Display,
span: &Span
) -> Self

For when the user defines an array range values that is >= the array length.

pub fn statement_conditional_boolean_fails_to_resolve_to_boolean(
actual: impl Display,
span: &Span
) -> Self

For when the statement conditional boolean fails to resolve to a boolean.

pub fn statement_indicator_calculation(name: impl Display, span: &Span) -> Self

For when there was an error in SnarkVM trying to do a bit and operation.

pub fn statement_invalid_number_of_definitions(
expected: impl Display,
actual: impl Display,
span: &Span
) -> Self

For when a multi definition statement found a differing number of values and variable names.

pub fn statement_multiple_definition(value: impl Display, span: &Span) -> Self

For when the user tries to assign multiple variables to a single value.

pub fn statement_multiple_returns(span: &Span) -> Self

For when a function returns multiple times.

pub fn statement_no_returns(expected: impl Display, span: &Span) -> Self

For when a function expects a return type and has no valid return statements.

pub fn statement_select_fail(
first: impl Display,
second: impl Display,
span: &Span
) -> Self

For when SnarkVM fails to conditionally select between values for a gadget.

pub fn statement_tuple_assign_index(span: &Span) -> Self

For when the user tries to index a single tuple more than once.

pub fn statement_tuple_assign_index_bounds(
index: impl Display,
length: impl Display,
span: &Span
) -> Self

For when the user tries to index a value from an tuple that is >= the tuple length.

pub fn statement_unassigned(span: &Span) -> Self

For when the user doesn’t assign or return the expression.

pub fn statement_undefined_variable(name: impl Display, span: &Span) -> Self

For when a statement tries to use an unknown variable.

pub fn statement_undefined_circuit_variable(
name: impl Display,
span: &Span
) -> Self

For when the user defines a statement that tries to access an unknown circuit member variable.

pub fn statement_loop_index_const(span: &Span) -> Self

For when the user uses a nont const value for an iteration range.

pub fn address_value_account_error(error: impl ErrorArg, span: &Span) -> Self

For when there is an issue with an address value account.

pub fn address_value_invalid_address(actual: impl Display, span: &Span) -> Self

For when there is an invalid address value.

pub fn address_value_missing_address(span: &Span) -> Self

For when an a address value was expected but none was found.

pub fn boolean_value_cannot_enforce(
operation: impl Display,
error: impl ErrorArg,
span: &Span
) -> Self

For when an a boolean operation cannot be enforced due to a SnarkVM synthesis error.

pub fn boolean_value_cannot_evaluate(
operation: impl Display,
span: &Span
) -> Self

For when an a invalid boolean operation is called.

pub fn boolean_value_invalid_boolean(actual: impl Display, span: &Span) -> Self

For when there is an invalid boolean value.

pub fn boolean_value_missing_boolean(
expected: impl Display,
span: &Span
) -> Self

For when an a boolean value was expected but none was found.

pub fn char_value_invalid_char(actual: impl Display, span: &Span) -> Self

For when there is an invalid char value.

pub fn field_value_negate_operation(error: impl ErrorArg, span: &Span) -> Self

For when negating a field value fails due to a SnarkVM synthesis error.

pub fn field_value_binary_operation(
operation: impl Display,
error: impl ErrorArg,
span: &Span
) -> Self

For when an a field operation cannot be enforced due to a SnarkVM synthesis error.

pub fn field_value_invalid_field(actual: impl Display, span: &Span) -> Self

For when there is an invalid field value.

pub fn field_value_missing_field(expected: impl Display, span: &Span) -> Self

For when an a field value was expected but none was found.

pub fn group_value_negate_operation(error: impl ErrorArg, span: &Span) -> Self

For when negating a group value fails due to a SnarkVM synthesis error.

pub fn group_value_binary_operation(
operation: impl Display,
error: impl ErrorArg,
span: &Span
) -> Self

For when an a group operation cannot be enforced due to a SnarkVM synthesis error.

pub fn group_value_invalid_group(actual: impl Display, span: &Span) -> Self

For when there is an invalid group value.

pub fn group_value_missing_group(expected: impl Display, span: &Span) -> Self

For when an a group value was expected but none was found.

pub fn group_value_synthesis_error(error: impl ErrorArg, span: &Span) -> Self

For when the synthesis of a group failed due to a SnarkVM synthesis error.

pub fn group_value_x_invalid(x: impl Display, span: &Span) -> Self

For when the x coordinate of a group is invalid.

pub fn group_value_y_invalid(y: impl Display, span: &Span) -> Self

For when the y coordinate of a group is invalid.

pub fn group_value_not_on_curve(element: impl Display, span: &Span) -> Self

For when the current group value is not on the current supported curve.

pub fn group_value_x_recover(span: &Span) -> Self

For when the x coordinate of a group could not be recovered.

pub fn group_value_y_recover(span: &Span) -> Self

For when the y coordinate of a group could not be recovered.

pub fn group_value_n_group(number: impl Display, span: &Span) -> Self

For when a group generator cannot be multiplied by some number.

pub fn integer_value_signed(error: impl ErrorArg, span: &Span) -> Self

For when an a signed integer operation cannot be enforced due to a SnarkVM synthesis error.

pub fn integer_value_unsigned(error: impl ErrorArg, span: &Span) -> Self

For when an a unsigned integer operation cannot be enforced due to a SnarkVM synthesis error.

pub fn integer_value_synthesis(error: impl ErrorArg, span: &Span) -> Self

For when an a integer operation cannot be enforced due to a SnarkVM synthesis error.

pub fn integer_value_negate_operation(span: &Span) -> Self

For when negating a integer value fails due to a SnarkVM synthesis error.

pub fn integer_value_binary_operation(
operation: impl Display,
span: &Span
) -> Self

For when an a binary integer operation cannot be enforced due to a SnarkVM synthesis error.

pub fn integer_value_integer_type_mismatch(
expected: impl Display,
received: impl Display,
span: &Span
) -> Self

For when there is an integer type mismatch, one kind was expected but another was received.

pub fn integer_value_invalid_integer(actual: impl Display, span: &Span) -> Self

For when there is an invalid integer value.

pub fn integer_value_missing_integer(
expected: impl Display,
span: &Span
) -> Self

For when an a integer value was expected but none was found.

pub fn integer_value_cannot_evaluate(
operation: impl Display,
span: &Span
) -> Self

For when an a integer operation has no implementation.

pub fn lengthof_can_only_be_used_on_arrays(span: &Span) -> Self

For when .len() method is used on non-array values/variables.

pub fn array_sizes_must_match_in_eq(
lhs: impl Display,
rhs: impl Display,
span: &Span
) -> Self

For when equality operator is used on arrays with different sizes.

pub fn num_exit_codes() -> i32