Module leo_errors::emitter

Contains traits and types for channels through which errors go.

Structs

Buffer

A buffer of Ts.

BufferEmitter

An Emitter that collects into a list.

Handler

A handler deals with errors and other compiler output.

HandlerInner 🔒

Contains the actual data for Handler. Modelled this way to afford an API using interior mutability.

StderrEmitter

A trivial Emitter using the standard error.

Traits

Emitter

Types that are sinks for compiler errors.

Type Definitions

ErrBuffer

A buffer of LeoErrors.

WarningBuffer

A buffer of LeoWarnings.