leo_errors

Enum PackageError

pub enum PackageError {
    Formatted(Formatted),
    Backtraced(Backtraced),
}

PackageError enum that represents all the errors for the leo-package crate.

Variants

Formatted(Formatted)

Backtraced(Backtraced)

Implementations

impl PackageError

pub fn failed_to_get_input_file_entry(error: impl ErrorArg) -> Self

For when getting a input file entry failed.

pub fn failed_to_get_input_file_type( file: impl Debug, error: impl ErrorArg, ) -> Self

For when getting the input file type failed.

pub fn invalid_input_file_type(file: impl Debug, type_: FileType) -> Self

For when getting the input file has an invalid file type.

pub fn failed_to_create_inputs_directory(error: impl ErrorArg) -> Self

For when creating the inputs directory failed.

pub fn failed_to_read_circuit_file(path: impl Debug) -> Self

For when reading the struct file failed.

pub fn failed_to_read_inputs_directory(error: impl ErrorArg) -> Self

For when reading the input directory failed.

pub fn failed_to_read_input_file(path: impl Debug) -> Self

For when reading the input file failed.

pub fn failed_to_read_snapshot_file(path: impl Debug) -> Self

For when reading the snapshot file failed.

pub fn failed_to_read_checksum_file(path: impl Debug) -> Self

For when reading the checksum file failed.

pub fn io_error_circuit_file(error: impl ErrorArg) -> Self

For when the struct file has an IO error.

pub fn io_error_checksum_file(error: impl ErrorArg) -> Self

For when the checksum file has an IO error.

pub fn io_error_main_file(error: impl ErrorArg) -> Self

For when the main file has an IO error.

pub fn failed_to_remove_circuit_file(path: impl Debug) -> Self

For when removing the struct file failed.

pub fn failed_to_remove_checksum_file(path: impl Debug) -> Self

For when removing the checksum file failed.

pub fn failed_to_remove_snapshot_file(path: impl Debug) -> Self

For when removing the snapshot file failed.

pub fn io_error_input_file(error: impl ErrorArg) -> Self

For when the input file has an IO error.

pub fn io_error_gitignore_file(error: impl ErrorArg) -> Self

For when the gitignore file has an IO error.

pub fn failed_to_create_source_directory( path: impl Display, error: impl ErrorArg, ) -> Self

For when creating the source directory failed.

pub fn failed_to_get_leo_file_entry(error: impl ErrorArg) -> Self

For when getting a Leo file entry failed.

pub fn failed_to_get_leo_file_extension(extension: impl Debug) -> Self

For when getting the source file extension failed.

pub fn invalid_leo_file_extension( file: impl Debug, extension: impl Debug, ) -> Self

For when the Leo file has an invalid extension.

pub fn failed_to_initialize_package( package: impl Display, path: impl Debug, error: impl Display, ) -> Self

For when the package failed to initialize.

pub fn invalid_package_name(package: impl Display) -> Self

For when the package has an invalid name.

pub fn directory_not_found(dirname: impl Display, path: impl Display) -> Self

For when opening a directory failed.

pub fn failed_to_create_directory( dirname: impl Display, error: impl ErrorArg, ) -> Self

For when creating a directory failed.

pub fn failed_to_remove_directory( dirname: impl Display, error: impl ErrorArg, ) -> Self

For when removing a directory failed.

pub fn failed_to_read_file(path: impl Display, error: impl Display) -> Self

For when file could not be read.

pub fn failed_to_get_file_name() -> Self

pub fn failed_to_set_cwd(dir: impl Display, error: impl ErrorArg) -> Self

pub fn failed_to_open_manifest(error: impl Display) -> Self

pub fn failed_to_read_manifest(error: impl Display) -> Self

pub fn failed_to_write_manifest(error: impl Display) -> Self

pub fn failed_to_create_manifest(error: impl Display) -> Self

pub fn failed_to_open_aleo_file(error: impl Display) -> Self

pub fn failed_to_create_aleo_file(error: impl Display) -> Self

pub fn failed_to_write_aleo_file(error: impl Display) -> Self

pub fn failed_to_remove_aleo_file(error: impl Display) -> Self

pub fn empty_source_directory() -> Self

pub fn source_directory_can_contain_only_one_file( directory: impl Display, ) -> Self

pub fn io_error_env_file(error: impl ErrorArg) -> Self

For when the environment file has an IO error.

pub fn failed_to_deserialize_manifest_file( path: impl Display, error: impl ErrorArg, ) -> Self

pub fn failed_to_serialize_manifest_file( path: impl Display, error: impl ErrorArg, ) -> Self

pub fn failed_to_deserialize_lock_file(error: impl ErrorArg) -> Self

pub fn invalid_lock_file_formatting() -> Self

pub fn unimplemented_command(command: impl Display) -> Self

pub fn invalid_file_name_dependency(name: impl Display) -> Self

pub fn dependency_not_found(name: impl Display) -> Self

pub fn conflicting_on_chain_program_name( first: impl Display, second: impl Display, ) -> Self

pub fn missing_on_chain_program_name() -> Self

pub fn failed_to_read_manifest_file( path: impl Display, error: impl ErrorArg, ) -> Self

pub fn insufficient_balance( address: impl Display, balance: impl Display, fee: impl Display, ) -> Self

pub fn execution_error(error: impl Display) -> Self

pub fn snarkvm_error(error: impl Display) -> Self

pub fn failed_to_load_package(path: impl Display) -> Self

pub fn conflicting_dependency(existing: impl Display, new: impl Display) -> Self

pub fn conflicting_manifest( expected_name: impl Display, manifest_name: impl Display, ) -> Self

pub fn invalid_network_name(name: impl Display) -> Self

pub fn failed_path(path: impl Display, err: impl Display) -> Self

pub fn num_exit_codes() -> i32