Enum leo_errors::errors::compiler::compiler_errors::CompilerError

pub enum CompilerError {
    Formatted(Formatted),
    Backtraced(Backtraced),
}

CompilerError enum that represents all the errors for the leo-compiler crate.

Variants

Formatted(Formatted)

Backtraced(Backtraced)

Implementations

impl CompilerError

pub fn file_read_error(path: impl Debug, error: impl ErrorArg) -> Self

For when the compiler can’t read a file from the provided path.

pub fn illegal_static_member_assignment(member: impl Display, span: Span) -> Self

For when a user tries to assign to a struct static member.

pub fn import_not_found(file_path: impl Display, span: Span) -> Self

pub fn cannot_open_cwd(err: impl ErrorArg, span: Span) -> Self

pub fn program_name_should_match_file_name(
program_name: impl Display,
file_name: impl Display,
span: Span
) -> Self

pub fn program_scope_name_does_not_match(
program_scope_name: impl Display,
file_name: impl Display,
span: Span
) -> Self

pub fn num_exit_codes() -> i32