Enum leo_errors::errors::flattener::flattener_errors::FlattenError

pub enum FlattenError {
    Formatted(Formatted),
    Backtraced(Backtraced),
}

CliError enum that represents all the errors for the leo-lang crate.

Variants

Formatted(Formatted)

Backtraced(Backtraced)

Implementations

impl FlattenError

pub fn binary_overflow(
left: impl Display,
op: impl Display,
right: impl Display,
right_type: impl Display,
span: Span
) -> Self

For when a constant operation would cause an overflow.

pub fn unary_overflow(left: impl Display, op: impl Display, span: Span) -> Self

For when a constant operation would cause an overflow.

pub fn loop_has_neg_value(value: impl Display, span: Span) -> Self

For when a loop uses a negative value.

pub fn num_exit_codes() -> i32