leo_errors::errors::parser::parser_warnings

Enum ParserWarning

pub enum ParserWarning {
    Formatted(Formatted),
    Backtraced(Backtraced),
}

ParserWarning enum that represents all the warnings for the leo-parser crate.

Variants

Formatted(Formatted)

Backtraced(Backtraced)

Implementations

impl ParserWarning

pub fn const_parameter_or_input(span: Span) -> Self

For when a user used const on a parameter or input instead of constant.

pub fn deprecated(keyword: impl Display, help: impl Display, span: Span) -> Self

For when a keyword is deprecated but could be used as a valid identifier.

pub fn num_exit_codes() -> i32