Enum leo_errors::package::package_errors::PackageError

pub enum PackageError {
    FormattedError(FormattedError),
    BacktracedError(BacktracedError),
}

PackageError enum that represents all the errors for the leo-package crate.

Variants

FormattedError(FormattedError)

BacktracedError(BacktracedError)

Implementations

impl PackageError

pub fn failed_to_create_imports_directory(error: impl ErrorArg) -> Self

For when creating the imports directory failed.

pub fn import_does_not_exist(package: impl Display) -> Self

For when the specified import does not exist.

pub fn failed_to_remove_imports_directory(error: impl ErrorArg) -> Self

For when removing the imports directory failed.

pub fn failed_to_create_inputs_directory(error: impl ErrorArg) -> Self

For when creating the inputs directory failed.

pub fn failed_to_get_input_file_entry(error: impl ErrorArg) -> Self

For when getting a input file entry failed.

pub fn failed_to_get_input_file_name(file: impl Debug) -> Self

For when getting the input file name failed.

pub fn failed_to_get_input_file_type(
file: impl Debug,
error: impl ErrorArg
) -> Self

For when getting the input file type failed.

pub fn invalid_input_file_type(file: impl Debug, type_: FileType) -> Self

For when getting the input file has an invalid file type.

pub fn failed_to_read_inputs_directory(error: impl ErrorArg) -> Self

For when reading the input directory failed.

pub fn failed_to_read_input_file(path: impl Debug) -> Self

For when reading the input file failed.

pub fn io_error_input_file(error: impl ErrorArg) -> Self

For when the input file has an IO error.

pub fn failed_to_read_state_file(path: impl Debug) -> Self

For when reading the state file failed.

pub fn io_error_state_file(error: impl ErrorArg) -> Self

For when the state file has an IO error.

pub fn failed_to_read_checksum_file(path: impl Debug) -> Self

For when reading the checksum file failed.

pub fn failed_to_remove_checksum_file(path: impl Debug) -> Self

For when removing the checksum file failed.

pub fn io_error_checksum_file(error: impl ErrorArg) -> Self

For when the checksum file has an IO error.

pub fn failed_to_read_circuit_file(path: impl Debug) -> Self

For when reading the circuit file failed.

pub fn failed_to_remove_circuit_file(path: impl Debug) -> Self

For when removing the circuit file failed.

pub fn io_error_circuit_file(error: impl ErrorArg) -> Self

For when the circuit file has an IO error.

pub fn failed_to_create_outputs_directory(error: impl ErrorArg) -> Self

For when creating the outputs directory failed.

pub fn failed_to_remove_outputs_directory(error: impl ErrorArg) -> Self

pub fn failed_to_read_proof_file(path: impl Debug) -> Self

For when reading the proof file failed.

pub fn failed_to_remove_proof_file(path: impl Debug) -> Self

For when removing the proof file failed.

pub fn io_error_proof_file(error: impl ErrorArg) -> Self

For when the proof file has an IO error.

pub fn failed_to_read_proving_key_file(path: impl Debug) -> Self

For when reading the prooving key failed.

pub fn failed_to_remove_proving_key_file(path: impl Debug) -> Self

For when removing the prooving key file failed.

pub fn io_error_proving_key_file(error: impl ErrorArg) -> Self

For when the proving key file has an IO error.

pub fn failed_to_read_verification_key_file(path: impl Debug) -> Self

For when reading the verification key file failed.

pub fn failed_to_remove_verification_key_file(path: impl Debug) -> Self

For when removing the verification key file failed.

pub fn io_error_verification_key_file(error: impl ErrorArg) -> Self

For when the verification key file has an IO error.

pub fn io_error_gitignore_file(error: impl ErrorArg) -> Self

For when the gitignore file has an IO error.

pub fn failed_to_create_manifest_file(
filename: impl Display,
error: impl ErrorArg
) -> Self

For when creating the manifest file failed.

pub fn failed_to_get_manifest_metadata_file(
filename: impl Display,
error: impl ErrorArg
) -> Self

For when getting the manifest file metadata failed.

pub fn failed_to_open_manifest_file(
filename: impl Display,
error: impl ErrorArg
) -> Self

For when opening the manifest file failed.

pub fn failed_to_parse_manifest_file(
filename: impl Display,
error: impl ErrorArg
) -> Self

For when parsing the manifest file failed.

pub fn failed_to_read_manifest_file(
filename: impl Display,
error: impl ErrorArg
) -> Self

For when reading the manifest file failed.

pub fn failed_to_write_manifest_file(
filename: impl Display,
error: impl ErrorArg
) -> Self

For when writing the manifest file failed.

pub fn io_error_manifest_file(error: impl ErrorArg) -> Self

For when the manifest file has an IO error.

pub fn io_error_readme_file(error: impl ErrorArg) -> Self

For when the readme file has an IO error.

pub fn failed_to_create_zip_file(error: impl ErrorArg) -> Self

For when creating the zip file failed.

pub fn failed_to_open_zip_file(error: impl ErrorArg) -> Self

For when opening the zip file failed.

pub fn failed_to_read_zip_file(error: impl ErrorArg) -> Self

For when reading the zip file failed.

pub fn failed_to_write_zip_file(error: impl ErrorArg) -> Self

For when writing the zip file failed.

pub fn failed_to_remove_zip_file(path: impl Debug) -> Self

For when removing the zip file failed.

pub fn failed_to_zip(error: impl ErrorArg) -> Self

For when zipping fails.

pub fn io_error_zip_file(error: impl ErrorArg) -> Self

For when the zip file has an IO error.

pub fn io_error_library_file(error: impl ErrorArg) -> Self

For when the library file has an IO error.

pub fn io_error_main_file(error: impl ErrorArg) -> Self

For when the main file has an IO error.

pub fn failed_to_create_source_directory(error: impl ErrorArg) -> Self

For when creating the source directory failed.

pub fn failed_to_get_source_file_entry(error: impl ErrorArg) -> Self

For when getting a source file entry failed.

pub fn failed_to_get_source_file_extension(extension: impl Debug) -> Self

For when getting the source file extension failed.

pub fn failed_to_get_source_file_type(
file: impl Debug,
error: impl ErrorArg
) -> Self

For when getting the source file type failed.

pub fn invalid_source_file_extension(
file: impl Debug,
extension: impl Debug
) -> Self

For when getting the source file has an invalid extension.

pub fn invalid_source_file_type(file: impl Debug, type_: FileType) -> Self

For when getting the source file has an invalid file type.

pub fn failed_to_read_source_directory(error: impl ErrorArg) -> Self

For when reading the source directory failed.

pub fn failed_to_initialize_package(
package: impl Display,
path: impl Debug
) -> Self

For when the package failed to initalize.

pub fn invalid_package_name(package: impl Display) -> Self

For when the package has an invalid name.