leo_errors::traits

Trait LeoMessageCode

pub trait LeoMessageCode: Sized {
    // Required methods
    fn exit_code(&self) -> i32;
    fn error_code(&self) -> String;
    fn warning_code(&self) -> String;
    fn code_mask() -> i32;
    fn message_type() -> String;
    fn is_error() -> bool;

    // Provided method
    fn code_identifier() -> i8 { ... }
}

MessageCode trait that all Errors should implement.

Required Methods

fn exit_code(&self) -> i32

Returns the error’s exit code for the program.

fn error_code(&self) -> String

Returns the prefixed error identifier.

fn warning_code(&self) -> String

Returns the prefixed warning identifier.

fn code_mask() -> i32

Returns the messages’s exit code mask, as to avoid conflicts.

fn message_type() -> String

Returns the message’s code type for the program.

fn is_error() -> bool

Returns if the message is an error or warning.

Provided Methods

fn code_identifier() -> i8

The LeoErrorCode which has a default code identifier of 037 (Leo upsidedown and backwards). This is to make the exit codes unique to Leo itself.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.