Struct leo_errors::common::formatted::Formatted

pub struct Formatted {
    pub span: Span,
    pub backtrace: Backtraced,
}

Formatted compiler error type undefined value x โ> file.leo: 2:8 | 2 | let a = x; | ^ | = help: Initialize a variable x first. Makes use of the same fields as a BacktracedError.

Fields
span: Span

The formatted error span information.

backtrace: Backtraced

The backtrace to track where the Leo error originated.

Implementations

impl Formatted

pub fn new_from_span<S>( message: S, help: Option<String>, code: i32, code_identifier: i8, type_: String, error: bool, span: Span, backtrace: Backtrace, ) -> Self
where S: ToString,

Creates a backtraced error from a span and a backtrace.

pub fn exit_code(&self) -> i32

Calls the backtraces error exit code.

pub fn error_code(&self) -> String

Returns an error identifier.

pub fn warning_code(&self) -> String

Returns an warning identifier.