Struct leo_errors::emitter::Handler

pub struct Handler {
    inner: RefCell<HandlerInner>,
}

A handler deals with errors and other compiler output.

Fields
inner: RefCell<HandlerInner>

The inner handler. RefCell is used here to avoid &mut all over the compiler.

Implementations

impl Handler

pub fn new(emitter: Box<dyn Emitter>) -> Self

Construct a Handler using the given emitter.

pub fn new_with_buf() -> (Self, BufferEmitter)

Construct a Handler that will append to buf.

pub fn with<T>( logic: impl for<'a> FnOnce(&'a Handler) -> Result<T, LeoError> ) -> Result<T, ErrBuffer>

Runs logic provided a handler that collects all errors into the String, or if there were none, returns some T.

pub fn emit_err<E: Into<LeoError>>(&self, err: E)

Emit the error err.

pub fn emit_warning(&self, warning: LeoWarning)

Emit the error err.

pub fn fatal_err(&self, err: LeoError) -> !

Emits the error err. This will immediately abort compilation.

pub fn err_count(&self) -> usize

The number of errors thus far.

pub fn warning_count(&self) -> usize

The number of warnings thus far.

pub fn had_errors(&self) -> bool

Did we have any errors thus far?

pub fn last_err(&self) -> Result<(), Box<LeoError>>

Gets the last emitted error’s exit code if it exists. Then exits the program with it if it did exist.

pub fn extend_if_error<T>(&self, res: Result<T, LeoError>) -> Result<T, ()>

Extend handler with error given res = Err(error).