leo_errors

Module errors

Contains the errors and warnings for the Leo language.

Re-exports

pub use self::ast::*;

pub use self::cli::*;

pub use self::compiler::*;

pub use self::flattener::*;

pub use self::loop_unroller::*;

pub use self::interpreter_halt::*;

pub use self::package::*;

pub use self::parser::*;

pub use self::static_analyzer::*;

pub use self::type_checker::*;

pub use self::name_validation::*;

pub use self::utils::*;

Modules

ast 🔒

Contains the AST error definitions.

cli 🔒

Contains the CLI error definitions.

compiler 🔒

Contains the Compiler error definitions.

flattener 🔒

Contains the Flattener error definitions.

interpreter_halt 🔒

loop_unroller 🔒

Contains the Loop Unroller error definitions. This module contains the loop unroller error definitions.

name_validation 🔒

Contains the Name Validation error definitions.

package 🔒

Contains the Package error definitions.

parser 🔒

Contains the Parser error definitions.

static_analyzer 🔒

Contains the Static Analyzer error definitions.

type_checker 🔒

Contains the Type Checker error definitions.

utils 🔒

Contains the Utils error definitions.

Enums

LeoError

The LeoError type that contains all sub error types. This allows a unified error type throughout the Leo crates.

LeoWarning

The LeoWarning type that contains all sub warning types. This allows a unified warning type throughout the Leo crates.

Type Aliases

Result

A global result type for all Leo crates, that defaults the errors to be a LeoError.