Module leo_errors::errors

Contains the errors and warnings for the Leo lang.

Re-exports

pub use self::ast::*;

pub use self::cli::*;

pub use self::compiler::*;

pub use self::flattener::*;

pub use self::input::*;

pub use self::loop_unroller::*;

pub use self::package::*;

pub use self::parser::*;

pub use self::type_checker::*;

Modules

ast

Contains the AST error definitions.

cli

Contains the CLI error definitions.

compiler

Contains the Compiler error definitions.

flattener

Contains the Flattener error definitions.

input

Contains the Input error definitions.

loop_unroller

package

Contains the Package error definitions.

parser

Contains the Parser error definitions.

type_checker

Contains the Type Checker error definitions.

Enums

LeoError

The LeoError type that contains all sub error types. This allows a unified error type throughout the Leo crates.

LeoWarning

The LeoWarning type that contains all sub error types. This allows a unified error type throughout the Leo crates.

Type Aliases

Result

A global result type for all Leo crates, that defaults the errors to be a LeoError.