Enum leo_errors::errors::type_checker::type_checker_error::TypeCheckerError

pub enum TypeCheckerError {
    Formatted(Formatted),
    Backtraced(Backtraced),
}

InputError enum that represents all the errors for the inputs part of leo-ast crate.

Variants

Formatted(Formatted)

Backtraced(Backtraced)

Implementations

impl TypeCheckerError

pub fn invalid_assignment_target(span: Span) -> Self

For when the parser encountered an invalid assignment target.

pub fn cannot_assign_to_const_input(input: impl Display, span: Span) -> Self

For when the user tries to assign to a const input.

pub fn cannot_assign_to_const_var(var: impl Display, span: Span) -> Self

For when the user tries to assign to a const input.

pub fn type_should_be(
type_: impl Display,
expected: impl Display,
span: Span
) -> Self

For when the user tries to assign to a const input.

pub fn could_not_determine_type(expr: impl Display, span: Span) -> Self

For when the type checker cannot determine the type of an expression.

pub fn unknown_sym(kind: impl Display, sym: impl Display, span: Span) -> Self

For when the user tries to return a unknown variable.

pub fn incorrect_num_args_to_call(
expected: impl Display,
received: impl Display,
span: Span
) -> Self

For when the user tries calls a function with the incorrect number of args.

pub fn expected_one_type_of(
expected: impl Display,
received: impl Display,
span: Span
) -> Self

For when one of the following types was expected.

pub fn invalid_int_value(
value: impl Display,
type_: impl Display,
span: Span
) -> Self

For when an integer is not in a valid range.

pub fn invalid_core_instruction(
circuit: impl Display,
function: impl Display,
span: Span
) -> Self

For when an invalid core instruction is used.

pub fn core_type_name_conflict(type_: impl Display, span: Span) -> Self

For when a circuit is created with the same name as a core type.

pub fn function_has_no_return(func: impl Display, span: Span) -> Self

For when a function doesn’t have a return statement.

pub fn incorrect_num_circuit_members(
expected: impl Display,
received: impl Display,
span: Span
) -> Self

For when the user tries initialize a circuit with the incorrect number of args.

pub fn missing_circuit_member(
circuit: impl Display,
member: impl Display,
span: Span
) -> Self

For when the user is missing a circuit member during initialization.

pub fn invalid_core_circuit_call(expr: impl Display, span: Span) -> Self

An invalid access call is made e.g., bool::MAX

pub fn duplicate_circuit_member(circuit: impl Display, span: Span) -> Self

Attempted to define more that one circuit member with the same name.

pub fn duplicate_record_variable(record: impl Display, span: Span) -> Self

Attempted to define more that one record variable with the same name.

pub fn undefined_type(type_: impl Display, span: Span) -> Self

Attempted to access an invalid circuit.

pub fn invalid_circuit_variable(
variable: impl Display,
circuit: impl Display,
span: Span
) -> Self

Attempted to access an invalid circuit variable.

pub fn required_record_variable(
name: impl Display,
type_: impl Display,
span: Span
) -> Self

pub fn record_var_wrong_type(
name: impl Display,
type_: impl Display,
span: Span
) -> Self

pub fn compare_address(operator: impl Display, span: Span) -> Self

pub fn incorrect_tuple_length(
expected: impl Display,
actual: impl Display,
span: Span
) -> Self

pub fn invalid_tuple(span: Span) -> Self

pub fn tuple_out_of_range(
index: impl Display,
length: impl Display,
span: Span
) -> Self

pub fn tuple_not_allowed(span: Span) -> Self

pub fn unreachable_code_after_return(span: Span) -> Self

pub fn loop_body_contains_return(span: Span) -> Self

pub fn unknown_annotation(annotation: impl Display, span: Span) -> Self

pub fn helper_function_inputs_cannot_have_modes(span: Span) -> Self

pub fn circuit_or_record_cannot_contain_record(
parent: impl Display,
child: impl Display,
span: Span
) -> Self

pub fn num_exit_codes() -> i32