Enum leo_errors::errors::type_checker::type_checker_error::TypeCheckerError

pub enum TypeCheckerError {
    Formatted(Formatted),
    Backtraced(Backtraced),
}

InputError enum that represents all the errors for the inputs part of leo-ast crate.

Variants

Formatted(Formatted)

Backtraced(Backtraced)

Implementations

impl TypeCheckerError

pub fn invalid_assignment_target(span: Span) -> Self

For when the parser encountered an invalid assignment target.

pub fn cannot_assign_to_const_input(input: impl Display, span: Span) -> Self

For when the user tries to assign to a const input.

pub fn cannot_assign_to_const_var(var: impl Display, span: Span) -> Self

For when the user tries to assign to a const input.

pub fn type_should_be( type_: impl Display, expected: impl Display, span: Span ) -> Self

For when the user tries to assign to a const input.

pub fn could_not_determine_type(expr: impl Display, span: Span) -> Self

For when the type checker cannot determine the type of an expression.

pub fn unknown_sym(kind: impl Display, sym: impl Display, span: Span) -> Self

For when the user tries to return a unknown variable.

pub fn incorrect_num_args_to_call( expected: impl Display, received: impl Display, span: Span ) -> Self

For when the user tries calls a function with the incorrect number of args.

pub fn expected_one_type_of( expected: impl Display, received: impl Display, span: Span ) -> Self

For when one of the following types was expected.

pub fn invalid_int_value( value: impl Display, type_: impl Display, span: Span ) -> Self

For when an integer is not in a valid range.

pub fn invalid_core_function( struct_: impl Display, function: impl Display, span: Span ) -> Self

For when an invalid core function is used.

pub fn core_type_name_conflict(type_: impl Display, span: Span) -> Self

For when a struct is created with the same name as a core type.

pub fn function_has_no_return(func: impl Display, span: Span) -> Self

For when a function doesn’t have a return statement.

pub fn incorrect_num_struct_members( expected: impl Display, received: impl Display, span: Span ) -> Self

For when the user tries initialize a struct with the incorrect number of args.

pub fn missing_struct_member( struct_: impl Display, member: impl Display, span: Span ) -> Self

For when the user is missing a struct member during initialization.

pub fn invalid_core_function_call(expr: impl Display, span: Span) -> Self

An invalid access call is made e.g., `SHA256::hash()

pub fn duplicate_struct_member(struct_: impl Display, span: Span) -> Self

Attempted to define more that one struct member with the same name.

pub fn duplicate_record_variable(record: impl Display, span: Span) -> Self

Attempted to define more that one record variable with the same name.

pub fn undefined_type(type_: impl Display, span: Span) -> Self

Attempted to access an invalid struct.

pub fn invalid_struct_variable( variable: impl Display, struct_: impl Display, span: Span ) -> Self

Attempted to access an invalid struct variable.

pub fn required_record_variable( name: impl Display, type_: impl Display, span: Span ) -> Self

pub fn record_var_wrong_type( name: impl Display, type_: impl Display, span: Span ) -> Self

pub fn compare_address(operator: impl Display, span: Span) -> Self

pub fn incorrect_tuple_length( expected: impl Display, actual: impl Display, span: Span ) -> Self

pub fn invalid_tuple(span: Span) -> Self

pub fn tuple_out_of_range( index: impl Display, length: impl Display, span: Span ) -> Self

pub fn unreachable_code_after_return(span: Span) -> Self

pub fn loop_body_contains_return(span: Span) -> Self

pub fn unknown_annotation(annotation: impl Display, span: Span) -> Self

pub fn regular_function_inputs_cannot_have_modes(span: Span) -> Self

pub fn struct_or_record_cannot_contain_record( parent: impl Display, child: impl Display, span: Span ) -> Self

pub fn invalid_mapping_type( component: impl Display, type_: impl Display, span: Span ) -> Self

pub fn only_transition_functions_can_have_finalize(span: Span) -> Self

pub fn finalize_input_mode_must_be_public(span: Span) -> Self

pub fn finalize_output_mode_must_be_public(span: Span) -> Self

pub fn finalize_in_finalize(span: Span) -> Self

pub fn invalid_operation_outside_finalize( operation: impl Display, span: Span ) -> Self

pub fn finalize_without_finalize_block(span: Span) -> Self

pub fn loop_body_contains_finalize(span: Span) -> Self

pub fn missing_return(span: Span) -> Self

pub fn finalize_block_must_not_be_empty(span: Span) -> Self

pub fn cannot_have_constant_output_mode(span: Span) -> Self

pub fn transition_function_inputs_cannot_be_const(span: Span) -> Self

pub fn incorrect_num_args_to_finalize( expected: impl Display, received: impl Display, span: Span ) -> Self

pub fn invalid_self_access(span: Span) -> Self

pub fn missing_finalize(span: Span) -> Self

pub fn finalize_name_mismatch( finalize_name: impl Display, function_name: impl Display, span: Span ) -> Self

pub fn invalid_type(type_: impl Display, span: Span) -> Self

pub fn can_only_call_inline_function(span: Span) -> Self

pub fn cannot_invoke_call_to_local_transition_function(span: Span) -> Self

pub fn loop_bound_must_be_a_literal(span: Span) -> Self

pub fn strings_are_not_supported(span: Span) -> Self

pub fn imported_program_cannot_import_program(span: Span) -> Self

pub fn too_many_transitions(max: impl Display, span: Span) -> Self

pub fn assign_unit_expression_to_variable(span: Span) -> Self

pub fn nested_tuple_type(span: Span) -> Self

pub fn composite_data_type_cannot_contain_tuple( data_type: impl Display, span: Span ) -> Self

pub fn function_cannot_take_tuple_as_input(span: Span) -> Self

pub fn finalize_cannot_take_tuple_as_input(span: Span) -> Self

pub fn nested_tuple_expression(span: Span) -> Self

pub fn finalize_statement_cannot_contain_tuples(span: Span) -> Self

pub fn expression_statement_must_be_function_call(span: Span) -> Self

pub fn lhs_tuple_element_must_be_an_identifier(span: Span) -> Self

pub fn lhs_must_be_identifier_or_tuple(span: Span) -> Self

pub fn unit_expression_only_in_return_statements(span: Span) -> Self

pub fn function_cannot_output_record(span: Span) -> Self

pub fn cyclic_struct_dependency(path: Vec<impl Display>) -> Self

pub fn cyclic_function_dependency(path: Vec<impl Display>) -> Self

pub fn struct_cannot_have_member_mode(span: Span) -> Self

pub fn cannot_call_external_inline_function(span: Span) -> Self

pub fn finalize_cannot_take_record_as_input(span: Span) -> Self

pub fn finalize_cannot_output_record(span: Span) -> Self

pub fn finalize_cannot_return_value(span: Span) -> Self

pub fn too_many_mappings(max: impl Display, span: Span) -> Self

pub fn invalid_associated_constant(expr: impl Display, span: Span) -> Self

A call to an invalid associated constant is made e.g., bool::MAX

pub fn invalid_core_constant( type_: impl Display, constant: impl Display, span: Span ) -> Self

For when an invalid core constant is called.

pub fn invalid_block_access(span: Span) -> Self

For when an invalid field of block is called.

pub fn invalid_operation_inside_finalize( operation: impl Display, span: Span ) -> Self

pub fn operation_must_be_in_finalize_block(span: Span) -> Self

pub fn loop_range_decreasing(span: Span) -> Self

pub fn loop_bound_type_mismatch(span: Span) -> Self

pub fn const_declaration_must_be_literal_or_tuple_of_literals( span: Span ) -> Self

pub fn loop_bound_must_be_literal_or_const(span: Span) -> Self

pub fn incorrect_num_tuple_elements( identifiers: impl Display, types: impl Display, span: Span ) -> Self

pub fn const_declaration_can_only_have_one_binding(span: Span) -> Self

pub fn array_empty(span: Span) -> Self

pub fn array_too_large( size: impl Display, max: impl Display, span: Span ) -> Self

pub fn array_element_cannot_be_tuple(span: Span) -> Self

pub fn array_element_cannot_be_record(span: Span) -> Self

pub fn num_exit_codes() -> i32